Add CSRF insert filter and filters to chain
Fixes #116
