+++ auth/src/test/scala/com/lookout/borderpatrol/auth/keymaster/KeymasterSpec.scala
@@ -511,103 +511,4 @@ class KeymasterSpec extends BorderPatrolSuite with MockitoSugar {
       Await.result(output)
     }
   }
-
-  behavior of "keymasterIdentityProviderChain"
-
-  it should "succeed and invoke the GET on loginManager" in {
-    val server = com.twitter.finagle.Httpx.serve(
-      "localhost:5678", mkTestService[Request, Response]{request =>
-        if (request.path.contains("check")) Response(Status.Ok).toFuture
-        else Response(Status.BadRequest).toFuture
-      })
-
-    try {
-      // Allocate and Session
-      val sessionId = sessionid.untagged
-
-      // Login manager request
-      val loginRequest = req("enterprise", "/check",
-        ("username" -> "foo"), ("password" -> "bar"))
-
-      // Original request
-      val origReq = req("enterprise", "/ent", ("fake" -> "drake"))
-      sessionStore.update[Request](Session(sessionId, origReq))
-
-      // Execute
-      val output = keymasterIdentityProviderChain(sessionStore).apply(
-        SessionIdRequest(ServiceRequest(loginRequest, cust1, one), sessionId))
-
-      // Validate
-      Await.result(output).status should be(Status.Ok)
-    } finally {
-      server.close()
-    }
-  }
-
-  it should "succeed and invoke the GET on identityManager" in {
-    val server = com.twitter.finagle.Httpx.serve(
-      "localhost:5678", mkTestService[Request, Response]{request =>
-        if (request.path.contains(keymasterIdManager.path.toString))
-          tap(Response(Status.Ok))(res => {
-            res.contentString = TokensEncoder(tokens).toString()
-            res.contentType = "application/json"
-          }).toFuture
-        else Response(Status.BadRequest).toFuture
-      })
-
-    try {
-      // Allocate and Session
-      val sessionId = sessionid.untagged
-
-      // Login manager request
-      val loginRequest = req("enterprise", "/loginConfirm",
-        ("username" -> "foo"), ("password" -> "bar"))
-
-      // Original request
-      val origReq = req("enterprise", "/ent", ("fake" -> "drake"))
-      sessionStore.update[Request](Session(sessionId, origReq))
-
-      // Execute
-      val output = keymasterIdentityProviderChain(sessionStore).apply(
-        SessionIdRequest(ServiceRequest(loginRequest, cust1, one), sessionId))
-
-      // Validate
-      Await.result(output).status should be(Status.Found)
-    } finally {
-      server.close()
-    }
-  }
-
-  behavior of "keymasterAccessIssuerChain"
-
-  it should "succeed and invoke the GET on accessManager" in {
-    val server = com.twitter.finagle.Httpx.serve(
-      "localhost:5678", mkTestService[Request, Response]{request =>
-        if (request.path.contains(keymasterAccessManager.path.toString))
-          tap(Response(Status.Ok))(res => {
-            res.contentString = TokensEncoder(tokens2).toString()
-            res.contentType = "application/json"
-          }).toFuture
-        else if (request.path.contains(one.path.toString)) Response(Status.Ok).toFuture
-        else Response(Status.BadRequest).toFuture
-      })
-
-    try {
-      // Allocate and Session
-      val sessionId = sessionid.untagged
-
-      // Original request
-      val origReq = req("enterprise", "/ent")
-      sessionStore.update[Tokens](Session(sessionId, tokens))
-
-      // Execute
-      val output = keymasterAccessIssuerChain(sessionStore).apply(
-        SessionIdRequest(ServiceRequest(origReq, cust1, one), sessionId))
-
-      // Validate
-      Await.result(output).status should be(Status.Ok)
-    } finally {
-      server.close()
-    }
-  }
 }
