+++ security/src/main/scala/com/lookout/borderpatrol/security/Csrf.scala
@@ -2,9 +2,9 @@ package com.lookout.borderpatrol.security
 
 import com.lookout.borderpatrol.util.Combinators.tap
 import com.lookout.borderpatrol.sessionx._
-import com.twitter.finagle.{Service, Filter}
-import com.twitter.finagle.httpx.{Request, Response}
-import com.twitter.util.Future
+import com.twitter.finagle.{SimpleFilter, Service, Filter}
+import com.twitter.finagle.httpx.{Status, Cookie, Request, Response}
+import com.twitter.util.{Time, Future}
 
 object Csrf {
   case class InHeader(val header: String = "X-BORDER-CSRF") extends AnyVal
@@ -48,10 +48,32 @@ object Csrf {
 }
 
 /**
+ * Inserts the CSRF cookie in a Response sent back to the client
+ *
+ * - It should be typically happen only once, perhaps after the successful login
+ */
+case class CsrfInsertFilter[A](cookieName: Csrf.CookieName)(implicit secretStore: SecretStoreApi)
+    extends Filter[A, Response, A, Response] {
+
+  def apply(req: A, service: Service[A, Response]): Future[Response] =
+    service(req).flatMap(res =>
+      for {
+        csrfSessionId <- SessionId.authenticated
+        cookie <- tap(new Cookie(cookieName.name, csrfSessionId.asBase64)) { cooki => {
+          cooki.path = "/"
+          cooki.httpOnly = true
+          cooki.maxAge = Time.now.until(csrfSessionId.expires)
+        }}.toFuture
+        _ <- res.addCookie(cookie).toFuture
+      } yield res
+    )
+}
+
+/**
  * Sets the CSRF header for inspection by upstream service. Always sets csrf verified header to false unless
  * the csrf cookie and the header/param match and are valid.
  */
-case class CsrfFilter(verify: Csrf.Verify) extends Filter[Request, Response, Request, Response] {
+case class CsrfVerifyFilter(verify: Csrf.Verify) extends SimpleFilter[Request, Response] {
 
   def apply(req: Request, service: Service[Request, Response]): Future[Response] =
     service(verify.unsafeInject(req)(_.toString))
