+++ server/src/main/scala/com/lookout/borderpatrol/server/services.scala
@@ -1,8 +1,13 @@
 package com.lookout.borderpatrol.server
 
+import com.lookout.borderpatrol.Binder.{ServiceIdentifierBinder, ManagerBinder, LoginManagerBinder}
+import com.lookout.borderpatrol.auth.OAuth2.OAuth2CodeVerify
 import com.lookout.borderpatrol.auth.keymaster.Keymaster._
 import com.lookout.borderpatrol.auth._
 import com.lookout.borderpatrol._
+import com.lookout.borderpatrol.auth.keymaster.{ServiceToken, Tokens}
+import com.lookout.borderpatrol.security.{CsrfVerifyFilter, CsrfInsertFilter}
+import com.lookout.borderpatrol.security.Csrf._
 import com.lookout.borderpatrol.sessionx.{SessionStore, SecretStoreApi}
 import com.twitter.finagle.Service
 import com.twitter.finagle.httpx.{Request, Response}
@@ -10,6 +15,32 @@ import com.twitter.finagle.stats.StatsReceiver
 
 
 object services {
+
+
+  /**
+   *  Keymaster Identity provider service Chain
+   * @param store
+   */
+  def keymasterIdentityProviderChain(store: SessionStore)(
+    implicit secretStoreApi: SecretStoreApi, statsReceiver: StatsReceiver): Service[SessionIdRequest, Response] =
+    LoginManagerFilter(LoginManagerBinder) andThen
+      KeymasterTransformFilter(new OAuth2CodeVerify) andThen
+      CsrfInsertFilter[KeymasterIdentifyReq](CookieName()) andThen
+      KeymasterPostLoginFilter(store) andThen
+      KeymasterIdentityProvider(ManagerBinder)
+
+
+  /**
+   * Keymaster Access Issuer service Chain
+   * @param store
+   */
+  def keymasterAccessIssuerChain(store: SessionStore)(
+    implicit secretStoreApi: SecretStoreApi, statsReceiver: StatsReceiver): Service[SessionIdRequest, Response] =
+    RewriteFilter() andThen
+      IdentityFilter[Tokens](store) andThen
+      AccessFilter[Tokens, ServiceToken](ServiceIdentifierBinder) andThen
+      KeymasterAccessIssuer(ManagerBinder, store)
+
 
   /**
    * Get IdentityProvider map of name -> Service chain
@@ -39,6 +70,7 @@ object services {
     val serviceMatcher = ServiceMatcher(config.customerIdentifiers, config.serviceIdentifiers)
 
     ExceptionFilter() andThen /* Convert exceptions to responses */
+      CsrfVerifyFilter(Verify(InHeader(), Param(), CookieName(), VerifiedHeader())) andThen /* Verify CSRF */
       ServiceFilter(serviceMatcher) andThen /* Validate that its our service */
       SessionIdFilter(config.sessionStore) andThen /* Get or allocate Session/SessionId */
       BorderService(identityProviderChainMap(config.sessionStore),
